[Bronze V] Title: A/B, Time: 92 ms, Memory: 12268 KB -BaekjoonHub

+++ 백준/Bronze/1008. A／B/A／B.kt
@@ -5,5 +5,5 @@ import java.util.StringTokenizer
 fun main() {
     val br = BufferedReader(InputStreamReader(System.`in`))
     val st = StringTokenizer(br.readLine())
-    println(st.nextToken().toDouble() / st.nextToken().toInt())
+    println(st.nextToken().toDouble() / st.nextToken().toDouble())
 }
